requirements.txt
appdirs==1.4.3
asn1crypto==0.23.0
beautifulsoup4==4.6.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
decorator==4.1.2
Django==1.11.7
djangorestframework==3.7.3
fudge==1.1.0
future==0.16.0
gmusicapi==10.1.2
gpsoauth==0.4.1
httplib2==0.10.3
idna==2.6
MechanicalSoup==0.8.0
mock==2.0.0
mutagen==1.38
oauth2client==4.1.2
pbr==3.1.1
pkg-resources==0.0.0
proboscis==1.2.6.0
protobuf==3.4.0
psycopg2==2.7.3.2
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
pycryptodomex==3.4.7
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
soundcloud==0.5.0
spotipy==2.4.4
urllib3==1.22
validictory==1.1.1
